+++ ui/src/logs/components/LogsTable.tsx
@@ -42,10 +42,13 @@ import {
   SeverityLevelColor,
   RowHeightHandler,
   SearchStatus,
+  Filter,
+  Operator,
 } from 'src/types/logs'
 import {INITIAL_LIMIT} from 'src/logs/actions'
 
 interface Props {
+  filters: Filter[]
   data: TableData
   isScrolledToTop: boolean
   isTruncated: boolean
@@ -660,7 +663,10 @@ class LogsTable extends Component<Props, State> {
   }
 
   private renderMessage = (formattedValue: string): JSX.Element => {
-    const {notify} = this.props
+    const {notify, filters} = this.props
+    const messageFilters = filters.filter(
+      f => f.key === 'message' && f.operator === Operator.LIKE
+    )
 
     if (this.props.isTruncated) {
       return (
@@ -668,11 +674,18 @@ class LogsTable extends Component<Props, State> {
           notify={notify}
           formattedValue={formattedValue}
           onExpand={this.props.onExpandMessage}
+          filters={messageFilters}
         />
       )
     }
 
-    return <LogsMessage notify={notify} formattedValue={formattedValue} />
+    return (
+      <LogsMessage
+        notify={notify}
+        formattedValue={formattedValue}
+        filters={messageFilters}
+      />
+    )
   }
 
   private severityDotStyle = (