+++ ui/src/logs/containers/LogsPage.tsx
@@ -237,6 +237,7 @@ class LogsPage extends Component<Props, State> {
               onExpandMessage={this.handleExpandMessage}
               notify={notify}
               searchStatus={searchStatus}
+              filters={filters}
             />
           </div>
         </div>