+++ ui/src/types/logs.ts
@@ -179,3 +179,14 @@ export enum Operator {
 }
 
 export type RowHeightHandler = (index: Index) => number
+
+export enum MatchType {
+  NONE = 'no-match',
+  MATCH = 'match',
+}
+
+export interface MatchSection {
+  id: string
+  type: MatchType
+  text: string
+}